% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MEDseq_news}
\alias{MEDseq_news}
\title{Show the NEWS file}
\usage{
MEDseq_news()
}
\value{
The \code{MEDseq} \code{NEWS} file, provided the session is interactive.
}
\description{
Show the \code{NEWS} file of the \code{MEDseq} package.
}
\examples{
MEDseq_news()
}
\keyword{utility}
