% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot.MEDseq}
\alias{plot.MEDseq}
\title{Plot MEDseq results}
\usage{
\method{plot}{MEDseq}(x,
       type = c("clusters", "mean", "precision", "gating", 
               "dbs", "asw", "bic", "icl", "aic", "nec", 
               "cv", "LOGLIK", "dbsvals", "aswvals", 
               "uncert.bar", "uncert.profile", "loglik", 
               "d", "f", "Ht", "i", "I"), 
       seriate = c("observations", "both", "clusters", "none"), 
       preczero = TRUE,
       log.scale = FALSE, 
       ...)
}
\arguments{
\item{x}{An object of class \code{"MEDseq"} generated by \code{\link{MEDseq_fit}} or an object of class \code{"MEDseqCompare"} generated by \code{\link{MEDseq_compare}}.}

\item{type}{A character string giving the type of plot requested:
\describe{
\item{"\code{clusters}"}{Visualise the data set with sequences grouped into their respective clusters. See \code{seriate}.}
\item{"\code{mean}"}{Visualise the central sequences. See \code{seriate}. The central sequence for the noise component, if any is not shown as it doesn't contribute in any way to the likelihood.}
\item{"\code{precision}"}{Visualise the central sequence parameters in the form of a heatmap. Values of \code{0} and \code{Inf} are shown in \code{grey} and \code{black} respectively (see \code{log.scale}).}
\item{"\code{gating}"}{Visualise the gating network, i.e. the observation index (by default) against the mixing proportions for that observation, coloured by cluster. See \code{seriate}. The optional argument \code{x.axis} can be passed via the \code{...} construct to change the x-axis against which mixing proportions are plotted.}
\item{"\code{dbs}"}{Plots all (weighted) mean/median DBS values in a fitted \code{MEDseq} object.}
\item{"\code{asw}"}{Plots all (weighted) mean/median ASW values in a fitted \code{MEDseq} object.}
\item{"\code{bic}"}{Plots all BIC values in a fitted \code{MEDseq} object.}
\item{"\code{icl}"}{Plots all ICL values in a fitted \code{MEDseq} object.}
\item{"\code{aic}"}{Plots all AIC values in a fitted \code{MEDseq} object.}
\item{"\code{nec}"}{Plots all NEC values in a fitted \code{MEDseq} object.}
\item{"\code{cv}"}{Plots all cross-validated log-likelihood values in a fitted \code{MEDseq} object.}
\item{"\code{LOGLIK}"}{Plots all maximal log-likelihood values in a fitted \code{MEDseq} object.}
\item{"\code{dbsvals}"}{Silhouette plot using observations-specific DBS values for the optimal model.}
\item{"\code{aswvals}"}{Silhouette plot using observations-specific ASW values for the optimal model.}
\item{"\code{uncert.bar}"}{Plot the observation-specific clustering uncertainties in the form of a bar plot.}
\item{"\code{uncert.profile}"}{Plot the observation-specific clustering uncertainties in the form of a profile plot.}
\item{"\code{loglik}"}{Plot the log-likelihood at every iteration of the EM/CEM algorithm used to fit the model.}
}
Also available are the following options which act as wrappers to types of plots produced by the \code{\link[TraMineR]{seqplot}} function in the \pkg{TraMineR} package.
\describe{
\item{"\code{d}"}{State distribution plots (by cluster).}
\item{"\code{f}"}{Sequence frequency plots (by cluster).}
\item{"\code{Ht}"}{Transversal entropy plots (by cluster).}
\item{"\code{i}"}{Selected sequence index plots (by cluster).}
\item{"\code{I}"}{Whole set index plots.}
}}

\item{seriate}{Switch indicating whether seriation should be used to improve the visualisation by re-ordering the \code{"observations"} within clusters (the default), the \code{"clusters"}, \code{"both"}, or \code{"none"}. See \code{\link[seriation]{seriate}}. The options \code{"clusters"} and \code{"both"} are only invoked when \code{type} is one of "\code{clusters}", "\code{mean}", "\code{precision}", "\code{gating}", "\code{d}", "\code{f}", "\code{Ht}", "\code{i}", or "\code{I}". Additionally, the options \code{"observations"} and \code{"both"} are only invoked when \code{type} is one of "\code{clusters}" or "\code{gating}".}

\item{preczero}{Logical indicating whether central sequence parameter positions corresponding to zero-valued precision parameters (if any!) should also be suppressed for the non-noise components. Defaults to \code{TRUE}; noise-component means are never shown regardless of the value of \code{preczero}.}

\item{log.scale}{Logical indicating whether precision parameter heatmaps should be plotted on the log-scale when \code{type="precision"}. The behaviour of \code{0} or \code{Inf} values remains unchanged; only strictly-positive finite entries are effected. Heavily imbalanced values are more likely for the "\code{UU}" and "\code{UUN}" model types, thus \code{log.scale} defaults to \code{TRUE} in those instances and \code{FALSE} otherwise.}

\item{...}{Catches unused arguments, and allows arguments to \code{\link{get_MEDseq_results}} to be passed when \code{type} is one of \code{"clusters"}, \code{"dbsvals"}, \code{"aswvals"}, \code{"uncert.bar"}, \code{"uncert.profile"}, \code{"d"}, \code{"f"}, \code{"Ht"}, \code{"i"}, or \code{"I"}, as well as the \code{x.axis} argument when \code{type="gating"}.}
}
\value{
The visualisation according to \code{type} of the results of a fitted \code{MEDseq} model.
}
\description{
Produces a range of plots of the results of fitted \code{MEDseq} models.
}
\details{
The \code{type} options related to model selection criteria plot values for \emph{all} fitted models in the "\code{MEDseq}" object \code{x}. The remaining \code{type} options plot results for the optimal model, by default. However, arguments to \code{get_MEDseq_results} can be passed via the \code{...} construct to plot corresponding results for suboptimal models in \code{x} when \code{type} is one of "\code{clusters}", "\code{d}", "\code{f}", "\code{Ht}", "\code{i}", or "\code{I}".
}
\note{
Every \code{type} of plot respects the sampling weights, if any. Those related to \code{\link[TraMineR]{seqdef}} plots from \pkg{TraMineR} may be too wide to display in the preview panel. The same is also true when \code{type} is "\code{dbsvals}" or "\code{aswvals}".
}
\examples{
\dontrun{
# Load the MVAD data
library(TraMineR)
data(mvad,   package = "TraMineR")
mvad$location <- factor(apply(mvad[,5:9], 1L, function(x) 
                 which(x == "yes")), labels = colnames(mvad[,5:9]))
mvad          <- list(sequences = mvad[,15L:86L], weights = mvad[,2], 
                      covariates = mvad[c(3:4,10:14,87)])
mvad.cov      <- mvad$covariates
states        <- c("EM", "FE", "HE", "JL", "SC", "TR")
labels        <- c("Employment", "FE", "HE", "Joblessness", "School", "Training")
mvad.seq      <- seqdef(mvad$sequences, states=states, labels=labels)
                        
# Fit a range of unweighted models without covariates
# Only consider models with a noise component
mod1          <- MEDseq_fit(mvad.seq, G=9:10, modtype=c("CCN", "CUN", "UCN", "UUN"))

# Plot the DBS values for all fitted models
plot(mod1, "dbs")

# Plot the clusters of the optimal model
plot(mod1, "clusters")

# Plot the clusters of the best UUN model
plot(mod1, "clusters", modtype="UUN")

# Fit a model with weights and gating covariates
# Drop the 1st sequence position which was used to define the weights
mvad.seq2     <- seqdef(mvad$sequences[,-1], states=states, labels=labels)
mod2          <- MEDseq_fit(mvad.seq2, G=10, modtype="UCN", weights=mvad$weights, 
                            gating=~ fmpr + gcse5eq + livboth, covars=mvad.cov)

# Plot the central sequences & precision parameters of this model
plot(mod2, "mean")
plot(mod2, "precision")

# Plot the clustering uncertainties in the form of a barplot
plot(mod2, "uncert.bar")

# Plot the observation-specific DBS values and the sequence frequencies by cluster
# Note that these plots may not display properly in the preview panel
plot(mod2, "dbsvals")
plot(mod2, "Ht")}
}
}
\seealso{
\code{\link{MEDseq_fit}}, \code{\link[TraMineR]{seqplot}}, \code{\link{dbs}}, \code{\link{get_MEDseq_results}}, \code{\link[seriation]{seriate}}
}
\author{
Keefe Murphy - <\email{keefe.murphy@ucd.ie}>
}
\keyword{main}
\keyword{plotting}
